Supporting MySql hit SQL_CALC_FOUND_ROWS for selecting row count. - refactoring
